Moved a struct only need by util into util
